refactor(model): model을 테이블 별로 정리하고 dto를 dto 패키지로 분리함
