+++ src/main/java/gift/authService/LoginService.java
@@ -1,9 +1,9 @@
 package gift.authService;
 
 import gift.exception.AuthException;
-import gift.model.Login;
-import gift.model.LoginRepository;
-import gift.model.Token;
+import gift.dto.Login;
+import gift.model.member.LoginRepository;
+import gift.dto.Token;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
