+++ src/main/java/gift/dto/Token.java
@@ -1,4 +1,4 @@
-package gift.model;
+package gift.dto;
 
 public record Token(String token) {
