+++ src/main/java/gift/model/wishlist/WishListRepository.java
@@ -1,5 +1,6 @@
-package gift.model;
+package gift.model.wishlist;
 
+import gift.dto.WishList;
 import java.util.List;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
